+++ core/src/test/java/com/syntaric/openfhir/mapping/kds/KdsGenericTest.java
@@ -3,4 +3,8 @@ package com.syntaric.openfhir.mapping.kds;
 import com.syntaric.openfhir.mapping.GenericTest;
 
 public abstract class KdsGenericTest extends GenericTest {
+    @Override
+    protected boolean containedToSeparateEntites() {
+        return false;
+    }
 }